% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangle.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Metodo para convertir nan's de data.frames a ceros}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
Metodo para convertir nan's de data.frames a ceros
}

