% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangle.R
\name{pcts}
\alias{pcts}
\title{Metodo para resumir porcentajes}
\usage{
pcts(v, base100 = TRUE)
}
\arguments{
\item{v}{vector de numeros}

\item{base100}{si TRUE cincuenta porciento es 50. Si FALSE multiplica por 100.}
}
\description{
Metodo para resumir porcentajes
}

