% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{t.i}
\alias{t.i}
\title{Importar varios archivos}
\usage{
\method{t}{i}(regex, ext)
}
\arguments{
\item{regex}{nombre de archivo o archivos en formato regex}

\item{ext}{nombre de extension con punto}
}
\description{
Metodo para importar uno o varios archivos. Regex y extension se ponen en dos parametros.
Regresa ademas un resumen.
}

