% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{t_actualizartodo}
\alias{t_actualizartodo}
\title{Actualizar todos los paquetes}
\usage{
t_actualizartodo()
}
\description{
Actualiza de github todos los paquetes de topazse.
Importante: se agregan dentro de la funcion de manera manual.
}

