% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangle.R
\name{t_eoutliers}
\alias{t_eoutliers}
\title{Eliminar outliers}
\usage{
t_eoutliers(.data, por, p = 0.01, tipo = "ambos")
}
\arguments{
\item{.data}{dataset a usar}

\item{por}{columna (desnuda) por la cual eliminar outliers}

\item{p}{porcentaje (en decimal) a eliminar de registros}

\item{tipo}{"ambos", "abajo" o "arriba" depende el caso}
}
\description{
Elimina outliers de columna al porcentaje parametrizado
}
\examples{
 df <- data.frame("columna1" = c(1,2,3,4,5,6,1,2,3,4,5,6),
                  "columna2" = c(84,102,99,103,101,109,
                                 116,121,122,119,131,222))
df_sin <- t_eoutliers(df, columna2)
}

