% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{t_i}
\alias{t_i}
\title{Importar varios archivos}
\usage{
t_i(regex, ext, mismas = FALSE)
}
\arguments{
\item{regex}{nombre de archivo o archivos en formato regex}

\item{ext}{nombre de extension con punto}

\item{mismas}{si TRUE, todos los archivos tienen el mismo formato (columnas). Si FALSE retorna solo el último.}
}
\description{
Metodo para importar uno o varios archivos. Regex y extension se ponen en dos parametros.
Regresa ademas un resumen.
}

