% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangle.R
\name{t_lu}
\alias{t_lu}
\title{Funcion para decirme rapidamente unicos por columna}
\usage{
t_lu(.data, col = 1, all = FALSE)
}
\arguments{
\item{.data}{data.frame}

\item{col}{numero de columa o nombre de columna}
}
\description{
Funcion para decirme rapidamente unicos por columna
}

