% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangle.R
\name{t_printnum}
\alias{t_printnum}
\title{Retorna strings de conteo}
\usage{
t_printnum(n, aprox = TRUE)
}
\arguments{
\item{n}{numero}

\item{aprox}{Si TRUE redondea, si FALSE pone todos los digitos.}
}
\description{
Al ingresar un numero, rendondea a miles o millones, de acuerdo al dato. Usado en prints para resumir informacion.
}

